poetrybot/requirements.txt
# Poetry Bot Pipeline - Requirements
# ===================================

# Core
python-dotenv>=1.0.0
pyyaml>=6.0
pydantic>=2.0

# Web Scraping
requests>=2.31.0
beautifulsoup4>=4.12.0
lxml>=4.9.0

# LLM / OpenAI API
openai>=1.0.0

# Image Generation
torch>=2.0.0
diffusers>=0.25.0
transformers>=4.36.0
accelerate>=0.25.0
safetensors>=0.4.0

# Text-to-Speech
edge-tts>=6.1.0
pydub>=0.25.1

# Audio/Video processing
ffmpeg-python>=0.2.0
soundfile>=0.12.0

# Scheduling
apscheduler>=3.10.0

# Testing
pytest>=7.4.0
pytest-asyncio>=0.21.0
